feat(agentLog): Add logging method and compression support for the agent
The node client should be able to serve as the transport mechanism
to the ingestion service for anything using nodejs.  LogDNA's agent
can be streamlined by re-using this package (and its retry logic)
to send data to the cloud.  This commit adds a method that should
only be used by the LogDNA agent, but supports the data structure
necessary for ingestion.  The added method also supports compression,
which is implemented with standard zlib.gzip compression.

Semver: minor
Ref: LOG-7103