fix: don't set 'user-agent' header when running in browser
Browsers do not allow you to set the 'user-agent' header and will block the
request if you attempt to do so.

This commit fixes this by first checking if we're running in a browser by
checking if the 'window' variable is defined, before attempting to set the
'user-agent' header.

See https://developer.mozilla.org/en-US/docs/Glossary/Forbidden_header_name for
more details on this behaviour.