feat: `ignoreRetryableErrors` will not emit retry-able errors
BREAKING CHANGE: The 'error' event's message was changed

Retry-able errors can get noisy considering they're
typically related to intermittent client/server errors
(currently being addressed), and there is no loss of data.
Rather than swallowing these errors, the user should be able
to control how noisy the errors are.  The default will be to
NOT emit retry-able errors.

To reduce confusion between irrecoverable errors, the `Error`
message was changed to reflect whether or not it's a temporary (or
retry-able) error.  Although minor, this is considered a breaking
change.

Semver: major
Ref: LOG-7639
Fixes: https://github.com/logdna/logger-node/issues/24