fix: Omit user-agent header if `sendUserAgent` option is `false`
When the logger is bundled for browser use, attempting to set the
user-agent can throw a console error of 'Refused to set unsafe header
"user-agent".'  Although the request is still sent, this makes for very
chattery console errors which is undesirable. Since not ALL browsers
print this error, the solution needs to be flexible enough to try
and set the user-agent for most cases, but also allow for it to be
turned off. The `sendUserAgent` constructor option, if set to
`false` will omit the user-agent header.  It is `true` by default.

Semver: minor
Ref: LOG-7793
Fixes: https://github.com/logdna/logger-node/issues/27