feat: Add more context (headers) to the errors
In some cases, HTTP errors can happen with things
like invalid header values. In order to properly debug
such situations, we need to report as much context
about the connection that we safely can. Part of this
will be exposing the header values. The URL of the
connection has also been requested in the past, and so
that has also been added to the error's metadata.

Semver: minor
Fixes: https://github.com/logdna/logger-node/issues/34