handle the situation when option property is falsy
When one of the properties of option is "" , has property returns true but it fails on typeOf check for it to be string. 
eg. value for env key is '', earlier has property returned true and it failed on checkStringParam. '' is string and should not have failed. So adding extra check for empty string on hasProperty function would help resolve this descrepency