refactor: extract _getPayloadSize method
* lib/logger.js: Extract logic for computing size of
a payload to a separate function so it can be overridden
or expanded cleanly (for example if we choose to start
measuring the size of the entire payload object instead
of just the length of the log message).

* test/logger-instantiation.js: Update tests relating to
count and set of methods in Logger.

No other test changes since _getPayloadSize is a trivial
internal utility function at this time.